Refactored flow to a separate file.
