+++ src/main.ts
@@ -11,13 +11,13 @@ import {
     ButtonComponent,
     Editor,
     TFile,
-    FuzzySuggestModal,
 } from "obsidian";
 
 import { ChoproFile } from "./parser";
 import { ChoproRenderer } from "./render";
 import { ChoproBlock } from "./parser";
 import { ChoproStyleManager } from "./styles";
+import { FlowGenerator } from "./flow";
 
 import {
     ChoproTransposer,
@@ -117,57 +117,8 @@ export default class ChoproPlugin extends Plugin {
     }
 
     private async openFlowFileSelector(editor: Editor) {
-        const modal = new FlowFileSelector(
-            this.app,
-            this.settings.flowFilesFolder,
-            async (file) => {
-                await this.insertFlowFromFile(file, editor);
-            }
-        );
-        modal.open();
-    }
-
-    private async insertFlowFromFile(file: TFile, editor: Editor) {
-        try {
-            const fileCache = this.app.metadataCache.getFileCache(file);
-            const flow = fileCache?.frontmatter?.flow;
-
-            if (!flow) {
-                new Notice("Selected file has no flow");
-                return;
-            }
-            
-            if (typeof flow === 'string') {
-                editor.replaceSelection(flow);
-                new Notice("Flow content inserted");
-
-            } else if (Array.isArray(flow)) {
-                const cursor = editor.getCursor();
-                let insertText = "";
-
-                for (const item of flow) {
-                    if (typeof item === 'string') {
-                        if (item.startsWith('#')) {
-                            // Section reference - create transclusion
-                            const sectionName = item.substring(1).trim();
-                            insertText += `![[${file.basename}${item}]]\n`;
-                        } else {
-                            // Markdown content
-                            insertText += `${item}\n`;
-                        }
-                    }
-                }
-
-                editor.replaceSelection(insertText.trim());
-                new Notice("Processed flow content");
-
-            } else {
-                new Notice("Flow property must be a string or array");
-            }
-        } catch (error) {
-            console.error("Error processing flow file:", error);
-            new Notice("Error processing flow file");
-        }
+        const flowGenerator = new FlowGenerator(this.app, this.settings);
+        await flowGenerator.openFlowFileSelector(editor);
     }
 
     async loadSettings() {
@@ -458,46 +409,5 @@ class TransposeModal extends Modal {
     onClose() {
         const { contentEl } = this;
         contentEl.empty();
-    }
-}
-
-class FlowFileSelector extends FuzzySuggestModal<TFile> {
-    private onSelect: (file: TFile) => Promise<void>;
-    private folderPath: string;
-
-    constructor(
-        app: App,
-        folderPath: string,
-        onSelect: (file: TFile) => Promise<void>
-    ) {
-        super(app);
-        this.onSelect = onSelect;
-        this.folderPath = folderPath;
-    }
-
-    getItems(): TFile[] {
-        let files = this.app.vault.getMarkdownFiles();
-        
-        if (this.folderPath && this.folderPath.trim() !== "") {
-            files = files.filter(file => 
-                file.path.startsWith(this.folderPath + "/") || 
-                file.path === this.folderPath
-            );
-        }
-        
-        // Only include files that have a flow property in frontmatter
-        return files.filter(file => {
-            const cache = this.app.metadataCache.getFileCache(file);
-            return cache?.frontmatter?.flow !== undefined;
-        });
-    }
-
-    onChooseItem(item: TFile, evt: MouseEvent | KeyboardEvent) {
-        this.close();
-        this.onSelect(item);
-    }
-
-    getItemText(item: TFile): string {
-        return item.path;
     }
 }
